+++ NEWS.md
@@ -5,6 +5,9 @@
 * `huber_loss()` is a numeric metric that is less sensitive to outliers than
 `rmse()`, but is more sensitive than `mae()` for small errors (@blairj09, #71).
 
+* `smape()` is a numeric metric that is based on percentage errors 
+(@riazhedayati, #67).
+
 # yardstick 0.0.2
 
 ## Breaking changes